requirements.txt
discord.py
requests
discord-py-buttons==1.1.0